semestral/requirements.txt
numpy
pygame
pillow
pytest
pylint
